move all nodewise markov tests from CheckMarkov Unit tests and allow direct call for nodewise markov
